JWT authentication: Allow placing additional constraints on tokens
As it is, we only validate the values of the "exp" (Expiration Time) and
"nbf" (Not Before) claims. This change extends the configuration file
schema to take a JMESPath expression, so that you can apply additional
constraints on tokens if needed.

More information about JMESPath can be found at https://jmespath.org/.