Unify the way authentication modules can provide metadata
Instead of allowing grpc.Authenticator to wrap context.Context any way
it wants, require those implementations to return a JSON-like value
(object, list or scalar), and place that value in the Context in a
unified way.

This makes it possible to provide a JMESPathExpressionAuthorizer that is
capable of making freeform authorization decisions based on the metadata
returned by the authentication code.