Replace buffer.WithBackgroundTask() with Buffer.WithTask()
The existing implementation of WithBackgroundTask() doesn't interact
nicely with WithErrorHandler(). The error handlers get pushed below the
background task, meaning that any errors returned by the background task
won't get caught by the error handler. We also don't wait for the
background task to complete in case the error handler triggers.

This change fixes this by letting tasks run in the foreground for
non-CAS buffers. This makes the implementation trivial. For CAS buffers,
we make the existing bufferWithBackgroundTask specific to the CAS. That
allows us to keep error handlers stacked on top of those buffers,
instead of pushing them below.

This fixes stacked usage of ReadFallbackBlobAccess.