Add ReadCanaryingBlobAccess
This backend makes it possible to do failover of REv2 storage traffic
from a local cache to the source of truth. This is useful in cases where
people run site-local caches in front of their cluster, and want builds
to continue while the site-local cache is down.

The backend works by defaulting to the main cluster, but doing periodic
probing to see whether the site-local cache is up. As soon as a single
request fails with an infrastructure error, it falls back to using the
cluster directly until the next successful probe.