Turn 'forward_and_reuse_metadata' into a global option
Right now there's a way to configure gRPC clients in such a way that
they forward metadata and also attach it to subsequent requests. This is
used by bb_clientd to ensure that reads against the FUSE file system
reuse the same credentials as the one provided by Bazel.

Instead of configuring it on a per-client basis, it makes more sense to
have a global option for it. This also ensures that credentials provided
by Bazel are used for all gRPC clients; not just the ones that requests
generated by Bazel end up using.

This is useful for when ReadCanaryingBlobAccess is used, as it means
that if bb_clientd switches between backends due to failover, it will
reuse the same credentials.