Replace hand-rolled concurrency with errgroup
We already use the errgroup type in a couple of places to introduce
bounded concurrency with proper cancellation. Two places where we don't
use it are MirroredBlobAccess and SizeDistinguishingBlobAccess. With
this change applied, we now properly cancel one outgoing request if the
other one fails.

While there, also fix a minor bug in MirroredBlobAccess where
FindMissing() may return NOT_FOUND errors. This can normally not happen;
only when a backend behaves inconsistently.Convert those to INTERNAL
errors, so the caller retries.