Add AnyAuthorizer
Just like how we have an AnyAuthenticator that sequentially calls into a
series of Authenticators until one returns success, this change adds an
Authorizer that does the same thing.

The intended use case for this is in bb_storage, where we want to allow
GetCapabilities() calls, only if the client is also allowed to either
read/write from the CAS/AC or execute actions. We want to combine
multiple Authorizers that are declared in config and turn them into a
single instance that guards GetCapabilities().