Add a new package for juggling with GetCapabilities() requests
In our case we want separate subsystems (storage, scheduler) to be able
to report their own capabilities and get them merged at a higher level.
This package adds the following:

- A Provider interface that various subsystems can implement to report
  their capabilities.
- A bunch of decorators for merging their responses and filtering their
  contents (authorization).
- A gRPC server type for the REv2 Capabilities service that calls into
  the Provider interface.

Now the goal is to adjust our subsystems to make use of the Provider
interface and patch up our tools to expose a REv2 Capabilities service
using the new server type.