Add a helper function for removing unsupported digest functions
If we're going to forward GetCapabilities() calls to remote backends and
respect the list of digest functions reported by the remote side, we
must also apply some filtering. It may be the case that the remote side
reports digest functions that we can't process ourselves. These need to
be removed if bb_storage is used as a forwarding proxy.