Add capabilities.Provider to BlobAccess
This change extends the API of BlobAccess; each implementation now has
to offer a GetCapabilities() method for reporting the capabilities of
the underlying backend. Only BlobAccess instances of storage type CAS
and AC need to implement it. For the ISCC and ICAS it's acceptable to
have an implementation that panics or crashes, as we won't let
bb_storage/main.go pass those interfaces to NewMergingProvider().

For backends that are primitive (i.e., local storage, HTTP, Redis), we
just embed an instance of capabilities.Provider. BlobAccessCreator is
repsonsible for providing a copy of the capabilities based on whether
the backend is used as a CAS or AC.

For sharding and mirroring we just forward the call to a single backend,
thereby assuming that all backends are configured equally. We may need
to revisit this in the future. For sharding it's likely too expensive to
query all backends, but for mirroring it may make sense to compute the
intersection of the capabilities to allow graceful upgrades.