Let BuildQueue use capabilities.Provider
Instead of letting implementations of BuildQueue implement a full
version of the GetCapabilities() RPC, we're going to trim it down. It
will now only be responsible for reporting ExecutionCapabilities.

Furthermore, we can remove types like UpdateEnabledTogglingBuildQueue
and NonExecutableBuildQueue. The former has been replaced by
AuthorizingProvider, while the latter is covered by MergingProvider.

We can then go ahead and patch up bb_storage, bb_scheduler and
bb_clientd to make use of the new capabilities framework in their main
source files. Notice that for bb_clientd this finally means we can drop
the dependency on the authorization package. This is correct, because
bb_clientd doesn't need to do any authorization checks; everything gets
forwarded from the remote side.

We do need to rework bb_storage a bit, because we now need to gather
capabilities providers from the CAS and AC BlobAccess instances as well
and mix those into the results. While there, structure this source file
to make it easier to eventually make the CAS and/or AC optional.

Fixes: #137