ACBlobAccessCreator: Make the CAS optional
Right now ACBlobAccessCreator can only be used to instantiate Action
Cache backends if a Content Addressable Storage is present as well. This
is because CompletenessCheckingBlobAccess depends on a CAS to work
properly.

This change removes this limitation by making the CAS parameter of
ACBlobAccess creator optional. We just add an explicit failure in case
one tries to use CompletenessCheckingBlobAccess when no CAS is provided.