Restructure the bb_storage configuration messages
Right now the storage authorizers are configured separately from the
storage backends themselves. This is confusing, as it means you can
specify authorizers, even if you don't want to use a backend of a given
type. Solve this by moving the BlobAccess and authorizer configurations
into a single message, making them siblings.

While there, remove the requirement that the CAS and AC are required.
Make it all consistent with how we configure the ICAS and ISCC. This
makes it possible to use Buildbarn as a CAS only data store, or to run
the CAS and AC on separate storage nodes.