pkg/digest: Add a utility function for encoding digests in binary form
For the NFSv4 server we need to encode digests in file handles that
identify objects in the file system. This change adds marshaling and
unmarshaling functions that store the hash and size in bytes in a
compact form.

The REv2 instance name is not stored in binary form, for the reason that
it can be arbitrarily long. We need to make sure that the size remains
bounded for it to be a valid NFSv4 file handle.