Get rid of FileTypeExecutableFile
I want to make use of the FileType enum in the virtual file system
package to act as FUSE/NFS independent definitions for S_IFREG, S_IFLNK,
etc.. The enum is an almost perfect fit, if it weren't for the fact that
executable files are distinct from regular files.

This change gets rid of this distinction. We now use a separate boolean
field in FileInfo to denote whether a file is executable, named
IsExecutable.