Overhaul the way we deal with device numbers
POSIX uses this type definition named dev_t that contains a block or
character device number. Traditionally this value is subdivided in a
major and minor number. The former represents the device class, with the
latter being the device instance in that class.

dev_t was originally 16 bits, meaning that the top 8 bits were used for
the major number and the bottom 8 bits for the minor number. This was of
course too small, meaning that each operating system made dev_t larger
and altered the major(), minor() and makedev() macros accordingly.
Unfortunately, they all use different arithmetic. golang.org/x/sys/unix
has methods for each operating system, and they all differ:

https://go.dev/src/cmd/vendor/golang.org/x/sys/unix/dev_darwin.go
https://go.dev/src/cmd/vendor/golang.org/x/sys/unix/dev_freebsd.go
https://go.dev/src/cmd/vendor/golang.org/x/sys/unix/dev_linux.go
https://go.dev/src/cmd/vendor/golang.org/x/sys/unix/dev_netbsd.go
https://go.dev/src/cmd/vendor/golang.org/x/sys/unix/dev_openbsd.go
https://go.dev/src/cmd/vendor/golang.org/x/sys/unix/dev_zos.go

What's worse is that not all of these functions are bijective, meaning
that a dev_t on one platform may not be representable as a major/minor
pair on another.

On modern operating systems that use a virtual device file system
(macOS, FreeBSD, etc.), the major/minor distinction is just a historical
remnant. On those systems, dev_t gets auto-generated. This is why I
patched up FreeBSD's ls(1) back in 2011 to no longer display device
numbers as a major/minor pair, but just a hexadecimal number:

https://github.com/freebsd/freebsd-src/commit/9f365aa1d6aab0aca269b54fb5d216ac0e31a06f

Back on topic: The reason for this change is that NFSv4 _only_ uses
major/minor numbers. There is no equivalent of a raw dev_t, where the
values are stored in composed form:

https://datatracker.ietf.org/doc/html/rfc7530#section-2.2.4

This means that we can no longer get away with just dealing with raw
dev_t values. We need to provide facilities to convert to both formats,
but also to do our best to not lose information in the process. This is
why I'm adding a composite type that stores device numbers in both
formats.

On Windows device numbers are non-existent. Provide a placeholder
implementation on that platform that only stores the major/minor pair.
This is sufficient for the only place where we will use them, namely the
NFSv4 server.