Implement graceful shutdown
This change lets all gRPC calls finish in case of SIGTERM or SIGINT.
This means that the client will not notice when the service is moving
around in e.g. Kubernetes. Additionally, the blob stores are flushed to
avoid loss of the latest data.

This change also introduces /-/ready to show when the gRPC sockets are listening.