Implement graceful termination of LocalBlobAccess
Right now LocalBlobAccess doesn't have any special code to do clean
shutdowns. This means that shutdowns may cause data loss. For many
production grade setups this is hardly an issue, because they use
mirrored centralized storage. There are, however, people who run
centralized storage setups that are either not mirrored or require
strong persistency. This change attempts to make LocalBlobAccess usable
for those setups as well.

First of all, this change adds a new InstallGracefulTerminationHandler()
method to pkg/global. That method creates a Context that callers can use
to detect that shutdown is taking place. Furthermore, it returns a
WaitGroup that things like LocalBlobAccess can use to delay the actual
shutdown.

We then patch up LocalBlobAccess to make use of the Context and
WaitGroup. Most notably, PeriodicSyncer's ProcessBlockPut() will unblock
itself if the Context gets cancelled. In that case it will perform a
double sync of the backing PersistentBlockList:

- Once to flush all of the data that was pending at the start of the
  termination.
- A second time to flush all of the data that written after termination
  was started.

The second time we pass in isFinalSync == true, causing the
PersistentBlockList to deny writes from that moment on. By only doing
that as part of the second sync, the actual amount of downtime is
minimised.

Partially fixes: https://github.com/buildbarn/bb-storage/pull/141
Patch by: Fredrik Medley