Add gRPC interceptor that copies message data to trace span attributes
In the past we've received PRs, where people started annotating trace
spans with attributes whose values were trivially obtained from RPC
message contents. I'm not a fan of those kinds of changes, for two
reasons:

- It litters the code with a bunch of boilerplate tracing code.
- What may be interesting for one, may not be of interest to the other.
  In other words, we blow up trace sizes, even though it may not be
  desirable.

To solve this, this change adds a new gRPC interceptor that can do what
we want in a configurable and automated manner. In config, we can
provide a map of RPC names, together with a list of fields we want to
extract.