Add ZIPReadingBlobAccess and ZIPWritingBlobAccess
I would like to write a utility for backing up actions stored in the CAS
and to restore them. That way actions can be inspected and rerun at a
later point in time. My plan is to just use ZIP files as a container
format for archived actions. Each file in the archive will represent a
single blob.

This change adds the code we need to be able to read archived actions
back in. It does this by adding a new BlobAccess backend that provides
read-only access to a ZIP archive. For files that are stored without
being compressed, it allows random file access. Oddly enough this is
supported by Go archive/zip package under the hood, but not exposed
through the API. The OpenRaw() method returns an io.Reader, even though
it should have likely returned an io.ReaderAt. We add a type assertion
to work around this limitation.

I'm not a big fan of doing this, but considering that we have testing
coverage for this code, any breakage in that area should be easy to
detect.

If we have a ZIP reader, we should also add a ZIP writer type. What's a
bit annoying about writing ZIP files is that the Go archive/zip package
isn't of too much use here. It's intended to be used to construct ZIP
files in a sequential manner, whereas BlobAccess is inherently parallel.

Let's solve this by writing our own ZIP files, according to the
specification given here:

https://pkware.cachefly.net/webdocs/APPNOTE/APPNOTE-6.3.9.TXT

The ZIP format doesn't support 64-bit files and offsets out of the box.
The archive/zip package adds ZIP64 headers whenever offsets/sizes start
to exceed 4 GB. Because we don't really care about storage size (read:
we don't even do compression), I've decided to just emit ZIP64 headers
unconditionally. This keeps the implementation simple.