Switch signal handling to an errgroup
I was mistaken. In bb-storage PR #141 I suggested _not_ using errgroup,
because we didn't care about propagating errors of shutting down storage
back to the signal handler. This was correct, if it weren't for the fact
that for bb_copy it would be most helpful to have the propagation.
Without it, main() can't detect that flushing storage failed, causing it
to return success erroneously.