Add NestedBlobReplicator
BlobReplicator on its own is only capable of replicating flat sets of
objects (stored in digest.Set). For bb_copy we need to have some logic
where we copy all of the transitive dependencies of CAS objects. Solve
this by introducing a helper type that can use any underlying
BlobReplicator to do the actual copying of data.