Add bb_copy: a generic blobstore copy utility
By now the blobstore layer has a very large number of BlobAccess storage
backends. We also have the BlobReplicator API that can be used to
replicate data from one backend to another.

One shortcoming is that BlobReplicator can now only be used in response
to REv2 storage traffic (e.g., part of MirroredBlobAccess). There is no
easy way to use BlobReplicator to do one-off copies of data. This change
attempts to address this by adding a simple command line tool for this.

There are various ways in which this tool can be used. For example,
copying an REv2 Directory including all of its children between two
clusters is as simple as invoking it with the following configuration:

    {
      source: { grpc: { address: 'cluster1.example.com:443' } },
      sink: { grpc: { address: 'cluster2.example.com:443' } },
      replicator: { 'local': {} },
      maximumMessageSizeBytes: 16 * 1024 * 1024,
      directories: [{
        hash: '3842e8454e054af62029d3bd2cb82d4ddeaf24a881c500e25440b54114f32dc2',
        sizeBytes: 710,
      }],
      traversalConcurrency: 10,
    }

With the recently added ZIPWritingBlobAccess backend you can also use it
to archive actions:

    {
      source: { grpc: { address: 'cluster1.example.com:443' } },
      sink: { zipWriting: { path: '/Users/ed/action.zip' } },
      replicator: { 'local': {} },
      maximumMessageSizeBytes: 16 * 1024 * 1024,
      directories: [{
        hash: '3842e8454e054af62029d3bd2cb82d4ddeaf24a881c500e25440b54114f32dc2',
        sizeBytes: 710,
      }],
      traversalConcurrency: 10,
    }

Conversely, restoring data from a ZIP file can be done using
ZIPReadingBlobAccess:

    {
      source: { zipReading: { path: '/Users/ed/action.zip' } },
      sink: { grpc: { address: 'cluster2.example.com:443' } },
      replicator: { 'local': {} },
      maximumMessageSizeBytes: 16 * 1024 * 1024,
      directories: [{
        hash: '3842e8454e054af62029d3bd2cb82d4ddeaf24a881c500e25440b54114f32dc2',
        sizeBytes: 710,
      }],
      traversalConcurrency: 10,
    }