Replace 'forwardMetadata' with a JMESPath expression
In some cases it may be of interest to set gRPC headers in a somewhat
programmatic fashion. Instead of just forwarding the literal value
provided to the client, it may be necessary to make slight alterations
to it. Or to set headers based on data yielded by the authentication
layer.

This change attempts to support such behaviour by removing the existing
'forwardMetadata' option, changing it to a single JMESPath expression
that has to yield a JSON object of type map[string][]string.

The change to grpc.proto includes information on how to convert uses of
the old 'forwardMetadata' option with 'addMetadataJmespathExpression'.