LocalBlobAccess: push down space allocation from BlockList into Block
Right now LocalBlobAccess has this restriction that objects are stored
aligned at sector boundaries. This is incredibly wasteful for small
objects (CAS Directory objects, AC ActionResults).

I would like to solve this by storing objects back to back, letting
BlockDeviceBackedBlockAllocator coordinate writes that happen close to
each other. This does require us to strengthen the interface of Block a
bit, as it currently permits writes to happen all over the place. The
API doesn't enforce that writes happen right after each other.

This change is does not actually address the issue at hand, but it at
least acts as a preparation for it. Once made, we can reimplement
BlockDeviceBackedBlockAllocator independent from the rest of the code to
follow a smarter allocation strategy.