Reimplement BlockDeviceBackedBlock.Put() to store data consecutively
With space allocation pushed down into implementations of Block, we can
improve BlockDeviceBackedBlock to store more than one object in a
sector. This change does exactly that.

BlockDevice.WriteAt() only permits calls that are sector aligned, having
a size that is multiple of the sector size. This means that whenever we
call WriteAt() on a sector that is shared between objects, we need to
ensure that that such writes are synchronized, and that subsequent
writes don't modify belonging to other objects. We solve this by adding
a sharedSector type that is tracked whenever a Block's write cursor is
not sector aligned.