Deal more gracefully with Tree objects that are too big
Workers currently don't have any hard limit on the size of the REv2 Tree
object they can emit. This means that we may store ActionResult entries
in the AC that reference objects that are too big.
CompletenessCheckingBlobAccess will then not be able to load those back,
causing GetActionResult() to fail with InvalidArgument. This leads to a
situation where clients get into a state where they can't make further
progress.

This change eliminates this situation by making it so that
CompletenessCheckingBlobAccess silently discards ActionResult entries
altogether, just like how it does that for ones that contain malformed
REv2 Digests.