Eliminate KeyBlobMap
Back when I created LocalBlobAccess, I added this KeyBlobMap layer. It's
basically a general purpose key-value store on which FlatBlobAccess was
built. When we added hierarchical instance names through
HierarchicalCASBlobAccess, we discovered that this abstraction wasn't a
perfect fit. That's why HierarchicalCASBlobAccess was written directly
on top of the underlying KeyLocationMap and LocationBlobMap.

Now that I want to add support for slicing REv2 Tree objects into
individual Directory objects, I've discovered that KeyBlobMap is also
not a great for FlatBlobAccess. It doesn't allow us to create new
KeyLocationMap entries for an already existing blob. Let's therefore
give FlatBlobAccess the same treatment as HierarchicalCASBlobAccess.

Small editorial changes have been performed to reduce the diff between
FlatBlobAccess and HierarchicalCASBlobAccess.