Add the BlobSlicer API
An interesting property of REv2 Tree objects is that if you look at
their binary representation, the Directory objects are contained in them
literally. This means that if LocalBlobAccess were to create
key-location map entries that point to subregions of the object, you can
quickly address individual Directory objects without loading the entire
Tree.

This change adds a new BlobAccess.GetFromComposite() method that can be
used to fetch a small object (Directory) that is contained within a
larger object (Tree). Implementations of BlobAccess will of course need
some smartness to figure out how to decompose the larger object into the
smaller ones. This is where the BlobSlicer interface comes in. It allows
the caller to provide arbitrary logic for decomposing.

This change is of course fairly large, as we need to patch up virtually
all implementations of BlobAccess to have a proper implementation.
Furthermore, we have a couple of BlobAccess implementations that call
into BlobReplicator. As a result, that interface also needs to gain a
ReplicateComposite() method.

Our main focus at this point is to have a high-quality implementation of
GetFromComposite() as part of LocalBlobAccess. For some of the other
backends we don't care that much, or it's simply infeasible to do
anything smart. In those cases we just forward the call to Get() and
invoke the BlobSlicer directly. This makes the method work as expected,
except that it doesn't end up retaining the decomposition.