Switch to the OpenTelemetry wire format for listing attributes
We currently have a custom Protobuf message for listing OpenTelemetry
trace attributes. It turns out that the OpenTelemetry gRPC wire format
also provides a message that is perfectly usable for this.

In an ideal world, the OpenTelemetry SDK for Go would provide a utility
function for doing the conversion from Protobuf to native for us.
Unfortunately, any code in this area seems to be off limits and is only
capable of converting from native to Protobuf; not the other way around.

This change adds a custom method for this purpose. We place it in
pkg/otel, so that we can easily test it. Furthermore, it can now be
reused by AuthenticationMetadata.