Add structure to the authentication metadata object
Right now authentication metadata is just a JSON-like object that is
generated by Authenticators, to be passed on to Authorizers. In our case
it would also be nice to also bolt on some other things:

- Trace attributes, so that gRPC trace spans contain information on the
  user performing the request.
- Information that is safe to display publicly (e.g., as part of logs,
  the bb_scheduler/bb_browser pages).

This change adds a Protobuf message named AuthenticationMetadata that
describes which fields in the authentication metadata object have a
special meaning. Furthermore, we add a native AuthenticationMetadata
type for it as well, which stores the same information in preprocessed
form. That way the metadata doesn't need to be computed for every
request.