Let authentication metadata be the Protobuf message
Right now we have a somewhat quirky scheme, where authentication
metadata is provided in the form of a JSON-like Go object. We use
protojson's DiscardUnknown to transform known fields such as "public"
and "tracingAttributes" into the Protobuf equivalent.

I think this is somewhat hard to understand for end users. Letting
certain fields have a special meaning is also not future proof, as it
means we can't always safely add new fields without risking breakage due
to overlaps.

Let's solve this by letting NewAuthenticationMetadata*() simply take the
Protobuf message. If private information needs to be tracked for use by
Authorizer, people can place it in the newly added "private" field.

This has the advantage that various configuration messages can now embed
buildbarn.auth.AuthenticationMetadata instead of a
google.protobuf.Value. It also means we get rid of some of the ambiguity
between map[string]any and map[string]string that we see in some of the
unit tests.