Postpone healthy service status response a bit
The healthy and serving status, i.e. HTTP /-/healthy and
grpc_health_v1.HealthCheckResponse_SERVING, will now be postponed until
the whole service is up and running. Before, the healthy status was
potentially reported before starting to listen to the gRPC ports.

This removes an error mode where Kubernetes would believe that the
service was ready just before it crashed. The previous behaviour could
lead to a cluster being upgraded to an erroneous state even for a
simple jsonnet configuration error.

Most of the diff in this commit is simply indentation change.