Add util.VisitProtoBytesFields()
My plan is to stop treating REv2 Tree objects are individual messages. I
would like to process them as if they are streams, containing a sequence
of REv2 Directory objects. In order to do that, we need to ship our very
own streaming Protobuf message parser, as Go's Protobuf library doesn't
ship anything like that out of the box.

This change adds a helper function for doing exactly that. It adds a
parser for messages that exclusively consist of bytes fields at the top
level.