Add field numbers of Protobuf message remoteexecution.Tree
In most of the places in the code base I want to stop parsing Tree
objects as if they are single Protobuf message. I want to parse them in
a streaming way, so that memory usage is kept to a minimum. For that we
need to know the field numbers of the 'root' and 'children' fields.
Let's add constants for those, copied from the REv2 spec.