Let CompletenessCheckingBlobAccess use util.VisitProtoBytesFields()
Now that we can process REv2 Tree objects in a streaming fashion, let's
reimplement CompletenessCheckingBlobAccess to make use of it. What's
interesting is that we can now treat the root and child directories
identically, as it's just the field number that's different.

What's slightly scary is that by streaming the Tree object, we're
processing the data speculatively. If it turns out the object read from
the CAS is corrupted, results to any of the calls to FindMissing() made
in response to that Tree should be suppressed. Only when the Tree object
has been read in its entirety is it safe to return the original error
response.

This change introduces a small change to the configuration schema. If
you previous had this:

    backend: { completenessChecking: ... },

You will now need to write something along these lines:

    backend: { completenessChecking: {
      backend: ...,
      maximumTotalTreeSizeBytes: 64 * 1024 * 1024,
    } },