Let Buildbarn binary entry points use structured concurrency
Right now we use global.InstallGracefulTerminationHandler() to install
signal handles for doing graceful shutdowns of applications. This system
is inadequate, as it cannot be used to express dependencies between
individual subsystems.

I want to add support for gracefully shutting down bb_worker, and in
that case we need to express dependencies to ensure BlobAccess and the
FUSE/NFSv4 mount are not shut down before the worker threads complete.
To solve that, I am adding a new program.Run() method to replace the
former.

program.Run() effectively behaves as an errgroup, but with a couple of
additions:

- Instead of returning an error at the very end (after all goroutines
  complete), it logs all of them as soon as they occur. This makes it
  easier to understand what's going on.

- Every routine spawned using program.Run() gets two program.Group
  handles that can be used to spawn additional routines. The first group
  can be used to spawn siblings, while the second can be used to spawn
  dependencies of the current goroutine and its siblings. The context of
  the dependencies will only be canceled if the current goroutine and
  its siblings terminate.

- Just like global.InstallGracefulTerminationHandler(), it installs a
  process wide signal handler.

One advantage of this mechanism is that we can now use `return err`
everywhere to initiate graceful shutdown. For example, if we cannot
spawn a gRPC server, we will still flush LocalBlobAccess before shutting
down.

Patch up places where we spawn gRPC/web servers to also launch their
goroutines as part of a program.Group. To prevent this from causing our
programs to hang during shutdown indefinitely, we need to let those
respect ctx. Solve this by spawning additional goroutines that monitor
<-ctx.Done().

There are still some places left where we spawn bare goroutines that
should now be created as part of program.Group. Add TODO comments to
indicate those locations.