LocalBlobAccess: Explicitly fail if the block sector count is zero
If LocalBlobAccess is backed by storage with a very large sector size,
it may be the case that the resulting sector count for each block rounds
down to zero. The downside of that is that you end up with a storage
backend that is incapable of storing any data.

We already had a small number of cases where people ran into this and
ended up asking questions on Slack. Let's make it easier for our users
to discover this on their own by adding a specific error message for it.