build: Modify patch to remote-asset-apis for greater compatibility
This change modifies the patch to the `bazelbuild/remote-apis` repo to
change the generated target name to one that happens to match the
existing target names more closely. This doesn't seem to make a
difference to `bb-storage`, but it does in repos like `remote-asset-api`
where the patch does not generate targets for the corresponding proto,
and therefore the deps of said proto and generated libraries must match
the ones generated in this patch.

A more robust fix might be to either:
* re-work the patch to cover each proto in a consistent manner
* upstream something to `remote-apis` that would obviate the need for a
  patch

...but this seems to me to be a minimal, low-risk change to get
everything working.

Tested: `bazel test //...` still passes on my system