LocalBlobAccess: Make LocationRecordArray prime sized
and HashingKeyLocationMap uses simple modulo arithmetic to pick a slot.
This means that if the LocationRecordArray has size 2^n, only a fraction
of the bits of the hash are used. Because we use FNV-1a, the lower bits
are known to be of poor quality. This means that we may see an elevated
number of hash collisions as a result.

Eliminate this problem entirely by ensuring that LocationRecordArray is
prime sized. This causes all bits of the hash to be taken into account.