Generalize program.Run()
program.Run()'s approach of placing goroutines in a tree to take
shutdown order into account seems to be pretty flexible. I would like to
be able to use a similar approach in more isolated settings, similar to
errgroup.Group.

This change decomposes the existing Run() function into two parts:

- run(), which calls into the main goroutine and waits for it and its
  descendants to complete.

- RunMain(), which does the signal handling, error logging and process
  shutdown.

With this decomposition in place, we can add a RunLocal() function that
can be used as a replacement for errgroup.