buffer: Prevent unnecessary calls against io.Reader
ChunkReader differs from a regular io.Reader in that it never returns
data and errors at the same time. This does however mean that if the
reader were to return an error (io.EOF or otherwise), it should be saved
and returned later.

Not all implementations of io.Reader (such as the HTTP client used by
the AWS SDK) are willing to hold on to errors for you.