Patch up ReadCanaryingBlobAccess to do automated retrying
Right now ReadCanaryingBlobAccess depends on higher levels of logic to
do retries in case accessing the replica backend fails. This is all
right for operations issued by Bazel (e.g., individual Bytestream
calls), but it is prohibitively expensive if it causes bb_worker to mark
actions as having failed.

This change extends ReadCanaryingBlobAccess to do the retrying against
the source backend itself.