Add authentication middleware to our HTTP servers
For our gRPC servers we already have the ability to configure
authentication policies. This makes it possible to do validation of
JWTs, TLS client certificates, etc.. This change adds a similar feature
for HTTP servers. For bb_storage this isn't very useful, because it only
exposes a diagnostics HTTP server. For bb_browser and bb_scheduler it is
more relevant, as those tend to receive user traffic.

In addition to a couple of basic authentication policies, like 'allow',
'deny' and 'any', we add 'jwt', which is backed by the same HTTP
Authorization header parser as the one we have for gRPC.

We also add one for OpenID Connect. I initially looked into using some
OpenID Connect libraries for Go, but I don't think any of them were
actually that good or even useful. This is why I decided to build it
directly on top of golang.org/x/oauth2. This library is perfect for what
we're doing, except that it doesn't allow easy testing due its
dependence on wall time. We solve this by applying a local change that
improves the existing facilities that are used by the tests of
golang.org/x/oauth2 itself.