Better support for OAuth 2.0 servers that don't return expiration times
RFC 6749 states that the "expires_in" field is optional. It's valid for
OAuth 2.0 servers to return token responses without specifying an
expiration time. We could add a configuration option to allow specifying
a default value in case the server doesn't report one. But instead of
doing that, let's use an exponential backoff.