Add Prometheus metrics to HTTP servers
We already have Prometheus metrics for HTTP clients (e.g., AWS SDK,
Pushgateway), but don't have it for servers. This has so far not been a
big issue because we tend to use gRPC more frequently, but now that we
have a more advanced authentication framework for HTTP, it makes sense
to give some more insight there as well.