Document how to tune the size of LocalBlobAccess
We often get questions on Slack how to do this. There are some
Prometheus metrics that can be inspected to determine whether the
key-location map and blocks backend are appropriately sized. Let's
document these explicitly in the Protobuf definitions for the storage
configuration.

While there, fix a typo in the metric name. It should be named
'old_current_new' to match the implementation, not 'old_new_current'.