Add HTTP authentication based on the "Accept" header
It may be desirable in certain cases to only perform OpenID Connect
based authentication in case users visit a page through the web browser.
For non-interactive API access it may be desirable to return HTTP 401
instead.

This change solves this by adding a decorator for Authenticator that can
conditionalize its invocation based on whether a supported media type is
provided.