ReferenceExpandingBlobAccess: return codes.Canceled (#1765)
We've observed that builds may sometimes fail with errors like these:

    Code 2: Failed to prefetch file "...": [...]: context canceled

This is caused by the fact that ReferenceExpandingBlobAccess may
sometimes return errors with codes.Unknown when requests are canceled.

Let's solve this by decorating the reader returned by the HTTP/S3/...
client, so that it transforms any errors to gRPC-style statuses. This
change should cause PrefetchingBuildExecutor to suppress returning these
errors, as it already explicitly checks for codes.Canceled.