JWT: Let the configuration take a JSON Web Key Set
Right now key material needs to be provided in the form of PEM/DER
format. Even though this is native in the sense that Go's crypto
libraries use it, it's not a common way of representing them in the
JWT ecosystem.

This change replaces the existing options with a new one named
'jwks_inline', which may hold a JWKS as specified in RFC 7517, chapter
five. The easiest way to convert existing public keys to JWKS is to
install the 'step' CLI and run the command below:

step crypto key format --jwk < mykey

https://smallstep.com/docs/step-cli/reference/crypto/key/format/

This work is based on a contribution by Morten Mjelva and Robert
Collins. Thanks a lot!

Fixes: #165
Fixes: #179