Remove deprecated storage backends
Ever since we published ADR #2 which introduced LocalBlobAccess,
ShardingBlobAccess and MirroredBlobAccess, the writing has been on the
wall that there would no longer be any place for backends such as Redis
and SizeDistinguishing.

The main reason I kept these two backends around, was because people in
the community were still interested in seeing S3 support reappear.
Because I have not received any serious proposal from members within the
community to do this in a way that it meets the standards for inclusion,
I think we've now reached the point where we can assume no work is going
to happen in this area in the short term.

Fixes: #175
Issue: #176